Add custom hooks
